Move TCPProxy to pkg
